fix: restore strict-zero analyzer coverage
