+++ env_inspector_core/__init__.py
@@ -1,8 +1,9 @@
 """Core engine for Env Inspector."""
 
 from __future__ import absolute_import, division
+from . import parsing
 from .cli import run_cli
 from .models import EnvRecord, OperationResult
 from .service import EnvInspectorService
 
-__all__ = ["run_cli", "EnvRecord", "OperationResult", "EnvInspectorService"]
+__all__ = ["run_cli", "EnvRecord", "OperationResult", "EnvInspectorService", "parsing"]
