+++ scripts/quality/__init__.py
@@ -1,0 +1,1 @@
+"""Quality gate scripts."""
