+++ scripts/quality/check_required_checks.py
@@ -4,11 +4,8 @@ from __future__ import absolute_import, division
 import argparse
 import importlib
 import json
-import os
 import sys
-import time
 import urllib.error
-from datetime import datetime, timezone
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Tuple
