+++ tests/test_resolver.py
@@ -28,6 +28,7 @@ def test_resolve_effective_windows_prefers_lower_precedence_rank():
         rec("powershell_profile", "windows", "API_TOKEN", "profile", 25),
     ]
     chosen = resolve_effective_value(rows, "API_TOKEN", "windows")
+    assert chosen is not None
     ensure(chosen is not None)
     ensure(chosen.value == "user")
 
@@ -38,6 +39,7 @@ def test_resolve_effective_wsl_context_isolated_by_distro():
         rec("wsl_bashrc", "wsl:Ubuntu", "API_TOKEN", "ubuntu-bash", 20),
     ]
     chosen = resolve_effective_value(rows, "API_TOKEN", "wsl:Ubuntu")
+    assert chosen is not None
     ensure(chosen is not None)
     ensure(chosen.value == "ubuntu-etc")
 
@@ -47,6 +49,7 @@ def test_resolve_effective_windows_does_not_leak_linux_context():
         rec("windows_user", "windows", "API_TOKEN", "windows-value", 20),
     ]
     chosen = resolve_effective_value(rows, "API_TOKEN", "windows")
+    assert chosen is not None
     ensure(chosen is not None)
     ensure(chosen.value == "windows-value")
 
@@ -58,5 +61,6 @@ def test_resolve_effective_linux_precedence_prefers_process_then_bashrc():
         rec("dotenv", "linux", "PATH", "dotenv", 90),
     ]
     chosen = resolve_effective_value(rows, "PATH", "linux")
+    assert chosen is not None
     ensure(chosen is not None)
     ensure(chosen.value == "process")
